+++ example-module.ts
@@ -1,11 +1,10 @@
 import type { Logger } from "tslog";
-import { Config, Override, BaseConfigurable } from "@msamblanet/node-config-types";
+import { Config, BaseConfigurable } from "@msamblanet/node-config-types";
 
 export interface FooConfig extends Config {
     a: number
     b: number
 }
-export type FooConfigOverride = Override<FooConfig>
 
 export class Foo extends BaseConfigurable<FooConfig> {
     public static readonly DEFAULT_CONFIG = { a: 1, b: 2 }