+++ HISTORY.md
@@ -2,5 +2,8 @@
 
 - UNRELEASED:
     - Expanded example, added template module
+    - Migrated to type-fest (from tsdef)
+    - Added mergeOptions
+    - Migrated to @msamblanet/node-project-template
 
 - 2021-09-08: Initial Release